enc importer: fix slur pixel-span heuristic in compound meters
Two bugs in resolvers-slur.cpp caused slurs to land on the wrong notes in
compound time signatures (e.g. 6/8):

1. startEncTick used beatTicks × timeSigDen instead of the correct
   durTicks × timeSigDen / timeSigNum (wt). In 6/8: 240×8=1920 vs 960.
   This shifted firstNoteXoff to the wrong note, making slur endpoints
   compute against the wrong reference.

2. endRel used the same wrong denominator (beatTicks × timeSigDen),
   producing a fraction smaller than the start tick and causing the slur
   to be dropped entirely.

Also treat SLURSTART xoffset as unsigned when computing pixel span:
values > 127 are stored as wrapped negative qint8 (e.g. 0x8A = 138px,
not -118) so signed arithmetic gives a spurious large span.

Regression test: importer_slur_pixel_span_6_8

+++ src/importexport/encore/internal/importer/noteloop-orn.cpp
@@ -156,8 +156,10 @@ void handleOrnament(BuildCtx& ctx, NoteLoopMeasCtx& mc, NoteElemCtx& ec)
         ps.startMeasIdx = measIdx;
         ps.endMeasIdx = endIdx;
         ps.alMezuro = static_cast<int>(eo->alMezuro);
-        ps.slurXoffset = static_cast<int>(eo->xoffset);
-        ps.slurXoffset2 = static_cast<int>(eo->xoffset2);
+        // Treat xoffset as unsigned: Encore stores pixel positions that wrap at 256.
+        // Values above 127 are stored as negative qint8 but represent large positive pixel offsets.
+        ps.slurXoffset  = static_cast<int>(static_cast<quint8>(eo->xoffset));
+        ps.slurXoffset2 = static_cast<int>(eo->xoffset2);  // already quint8
         ps.staffIdx = staffIdx;
         ps.encVoice = voice;
         ctx.pendingSlurs.push_back(ps);
